+++ ocnus/src/obser/vector.rs
@@ -16,7 +16,7 @@ use std::{
 pub struct ObserVec<const N: usize>(
     #[into_iterator(owned, ref, ref_mut)]
     #[serde(with = "serde_arrays")]
-    pub [f32; N],
+    pub [f64; N],
 );
 
 impl<const N: usize> ObserVec<N> {
@@ -31,25 +31,25 @@ impl<const N: usize> ObserVec<N> {
     }
 
     /// Calculate the mean square error between two observation vectors.
-    pub fn mse(&self, other: &Self) -> f32 {
-        (self - other).ss() / N as f32
+    pub fn mse(&self, other: &Self) -> f64 {
+        (self - other).ss() / N as f64
     }
 
     /// Calculate the mean square error between an observation vector
     /// and an optional reference observation.
-    pub fn mse_ref(&self, other_opt: Option<&Self>) -> f32 {
+    pub fn mse_ref(&self, other_opt: Option<&Self>) -> f64 {
         if self.any_nan() & other_opt.unwrap_or(&ObserVec::default()).any_nan() {
             0.0
         } else if !self.any_nan() & !other_opt.unwrap_or(&ObserVec::default()).any_nan() {
-            (self - other_opt.unwrap()).ss() / N as f32
+            (self - other_opt.unwrap()).ss() / N as f64
         } else {
-            f32::NAN
+            f64::NAN
         }
     }
 
     /// Calculate the sum of squares over the entries within the observation vector.
-    pub fn ss(&self) -> f32 {
-        self.iter().map(|value| value.powi(2)).sum::<f32>()
+    pub fn ss(&self) -> f64 {
+        self.iter().map(|value| value.powi(2)).sum::<f64>()
     }
 
     /// Return a new observation vector filled with zeros.
@@ -65,7 +65,7 @@ impl<const N: usize> Add for ObserVec<N> {
         ObserVec::<N>(
             zip_eq(self, rhs)
                 .map(|(a, b)| a + b)
-                .collect::<Vec<f32>>()
+                .collect::<Vec<f64>>()
                 .try_into()
                 .unwrap(),
         )
@@ -79,7 +79,7 @@ impl<'a, const N: usize> Add<&'a ObserVec<N>> for &'a ObserVec<N> {
         ObserVec::<N>(
             zip_eq(self, rhs)
                 .map(|(v1, v2)| v1 + v2)
-                .collect::<Vec<f32>>()
+                .collect::<Vec<f64>>()
                 .try_into()
                 .unwrap(),
         )
@@ -94,7 +94,7 @@ impl<const N: usize> AddAssign for ObserVec<N> {
 
 impl<const N: usize> Default for ObserVec<N> {
     fn default() -> Self {
-        ObserVec([f32::nan(); N])
+        ObserVec([f64::nan(); N])
     }
 }
 
@@ -114,62 +114,62 @@ impl Display for ObserVec<4> {
     }
 }
 
-impl<const N: usize> From<SVector<f32, N>> for ObserVec<N> {
-    fn from(value: SVector<f32, N>) -> Self {
+impl<const N: usize> From<SVector<f64, N>> for ObserVec<N> {
+    fn from(value: SVector<f64, N>) -> Self {
         ObserVec::from(value.data.0[0])
     }
 }
 
-impl<const N: usize> Mul<f32> for ObserVec<N> {
+impl<const N: usize> Mul<f64> for ObserVec<N> {
     type Output = ObserVec<N>;
 
-    fn mul(self, rhs: f32) -> Self::Output {
+    fn mul(self, rhs: f64) -> Self::Output {
         ObserVec::<N>(
             self.iter()
                 .map(|v| *v * rhs)
-                .collect::<Vec<f32>>()
+                .collect::<Vec<f64>>()
                 .try_into()
                 .unwrap(),
         )
     }
 }
 
-impl<const N: usize> Mul<f32> for &ObserVec<N> {
+impl<const N: usize> Mul<f64> for &ObserVec<N> {
     type Output = ObserVec<N>;
 
-    fn mul(self, rhs: f32) -> Self::Output {
+    fn mul(self, rhs: f64) -> Self::Output {
         ObserVec::<N>(
             self.iter()
                 .map(|v| *v * rhs)
-                .collect::<Vec<f32>>()
+                .collect::<Vec<f64>>()
                 .try_into()
                 .unwrap(),
         )
     }
 }
 
-impl<const N: usize> Mul<ObserVec<N>> for f32 {
+impl<const N: usize> Mul<ObserVec<N>> for f64 {
     type Output = ObserVec<N>;
 
     fn mul(self, rhs: ObserVec<N>) -> Self::Output {
         ObserVec::<N>(
             rhs.iter()
                 .map(|v| *v * self)
-                .collect::<Vec<f32>>()
+                .collect::<Vec<f64>>()
                 .try_into()
                 .unwrap(),
         )
     }
 }
 
-impl<'a, const N: usize> Mul<&'a ObserVec<N>> for f32 {
+impl<'a, const N: usize> Mul<&'a ObserVec<N>> for f64 {
     type Output = ObserVec<N>;
 
     fn mul(self, rhs: &'a ObserVec<N>) -> Self::Output {
         ObserVec::<N>(
             rhs.iter()
                 .map(|v| *v * self)
-                .collect::<Vec<f32>>()
+                .collect::<Vec<f64>>()
                 .try_into()
                 .unwrap(),
         )
@@ -194,7 +194,7 @@ impl<const N: usize> Sub for ObserVec<N> {
         ObserVec::<N>(
             zip_eq(self, rhs)
                 .map(|(v1, v2)| v1 - v2)
-                .collect::<Vec<f32>>()
+                .collect::<Vec<f64>>()
                 .try_into()
                 .unwrap(),
         )
@@ -214,7 +214,7 @@ impl<'a, const N: usize> Sub<&'a ObserVec<N>> for &'a ObserVec<N> {
         ObserVec::<N>(
             zip_eq(self, rhs)
                 .map(|(v1, v2)| *v1 - *v2)
-                .collect::<Vec<f32>>()
+                .collect::<Vec<f64>>()
                 .try_into()
                 .unwrap(),
         )
@@ -269,13 +269,13 @@ mod tests {
         assert!(ObserVec::<3>::default().any_nan());
         assert!(ObserVec::<3>::default().is_nan());
 
-        assert!(ObserVec([f32::NAN, 0.0, 0.0]).any_nan());
-        assert!(!ObserVec([f32::NAN, 0.0, 0.0]).is_nan());
+        assert!(ObserVec([f64::NAN, 0.0, 0.0]).any_nan());
+        assert!(!ObserVec([f64::NAN, 0.0, 0.0]).is_nan());
 
-        assert!(ObserVec([0.0, f32::NAN, 0.0, 0.0]).any_nan());
-        assert!(!ObserVec([0.0, f32::NAN, 0.0, 0.0]).is_nan());
+        assert!(ObserVec([0.0, f64::NAN, 0.0, 0.0]).any_nan());
+        assert!(!ObserVec([0.0, f64::NAN, 0.0, 0.0]).is_nan());
 
-        assert!(ObserVec([0.0, f32::NAN, 0.0, 0.0]) != ObserVec([0.0, f32::NAN, 0.0, 0.0]));
+        assert!(ObserVec([0.0, f64::NAN, 0.0, 0.0]) != ObserVec([0.0, f64::NAN, 0.0, 0.0]));
         assert!(ObserVec::<4>::zeros() == ObserVec::<4>::zeros());
 
         assert!(ObserVec::<4>::default().is_nan());
