+++ ocnus/src/stats/covmat.rs
@@ -12,27 +12,27 @@ use std::ops::{Mul, MulAssign};
 #[derive(Clone, Debug, Deref, Deserialize, Serialize)]
 pub struct CovMatrix {
     /// The  lower triangular matrix from the Cholesky decomposition.
-    cholesky_ltm: DMatrix<f32>,
+    cholesky_ltm: DMatrix<f64>,
 
     /// The inverse of the covariance matrix.
-    inverse_matrix: DMatrix<f32>,
+    inverse_matrix: DMatrix<f64>,
 
     /// The underlying dynamically sized covariance matrix.
     #[deref]
-    matrix: DMatrix<f32>,
+    matrix: DMatrix<f64>,
 
     /// The pseudo-determinant of the covariance matrix.
-    pseudo_determinant: f32,
+    pseudo_determinant: f64,
 }
 
 impl CovMatrix {
     /// Returns a reference to the lower triangular matrix L from the Cholesky decomposition.
-    pub fn cholesky_ltm(&self) -> &DMatrix<f32> {
+    pub fn cholesky_ltm(&self) -> &DMatrix<f64> {
         &self.cholesky_ltm
     }
 
     /// Create a [`CovMatrix`] from a semi positive-definite square matrix.
-    pub fn from_matrix(matrix: &DMatrixView<f32>) -> Result<Self, StatsError> {
+    pub fn from_matrix(matrix: &DMatrixView<f64>) -> Result<Self, StatsError> {
         let matrix_owned = matrix.into_owned();
 
         let (cholesky_ltm, determinant) = match matrix.cholesky() {
@@ -73,8 +73,8 @@ impl CovMatrix {
 
     /// Create a [`CovMatrix`] from a matrix of D-dimensional vectors.
     pub fn from_vectors<const D: usize>(
-        vectors: &MatrixView<f32, Const<D>, Dyn>,
-        opt_weights: Option<&[f32]>,
+        vectors: &MatrixView<f64, Const<D>, Dyn>,
+        opt_weights: Option<&[f64]>,
     ) -> Result<Self, StatsError> {
         let mut matrix = DMatrix::from_iterator(
             D,
@@ -98,7 +98,7 @@ impl CovMatrix {
         );
 
         // Fill up the other side of the matrix.
-        matrix += matrix.transpose() - DMatrix::<f32>::from_diagonal(&matrix.diagonal());
+        matrix += matrix.transpose() - DMatrix::<f64>::from_diagonal(&matrix.diagonal());
 
         // Detect columns/rows that need to be modified and modify them.
         let zero_variance_indices = (0..vectors.nrows())
@@ -137,24 +137,24 @@ impl CovMatrix {
     }
 
     /// Returns a reference to the inverse of the covariance matrix.
-    pub fn inverse_matrix(&self) -> &DMatrix<f32> {
+    pub fn inverse_matrix(&self) -> &DMatrix<f64> {
         &self.inverse_matrix
     }
 
     /// Returns a reference to the covariance matrix.
-    pub fn matrix(&self) -> &DMatrix<f32> {
+    pub fn matrix(&self) -> &DMatrix<f64> {
         &self.matrix
     }
 
-    /// Compute the multivariate likelihood from two iterators over `f32`.
+    /// Compute the multivariate likelihood from two iterators over `f64`.
     /// The length of both iterators must be equal and also a multiple
     /// of the dimension of the covariance matrix (panic).
     pub fn multivariate_likelihood(
         &self,
-        x: impl IntoIterator<Item = f32>,
-        mu: impl IntoIterator<Item = f32>,
-    ) -> f32 {
-        let delta = DVector::from(zip_eq(x, mu).map(|(i, j)| i - j).collect::<Vec<f32>>());
+        x: impl IntoIterator<Item = f64>,
+        mu: impl IntoIterator<Item = f64>,
+    ) -> f64 {
+        let delta = DVector::from(zip_eq(x, mu).map(|(i, j)| i - j).collect::<Vec<f64>>());
 
         let ndim = delta.len() / self.ndim();
 
@@ -163,7 +163,7 @@ impl CovMatrix {
         }
 
         let mut lh =
-            -(self.pseudo_determinant.ln() + ndim as f32 * std::f32::consts::PI * 2.0) / 2.0;
+            -(self.pseudo_determinant.ln() + ndim as f64 * std::f64::consts::PI * 2.0) / 2.0;
 
         for idx in 0..ndim {
             let view = delta.rows_with_step(idx, self.ndim(), ndim - 1);
@@ -185,15 +185,15 @@ impl CovMatrix {
     }
 
     /// Returns the pseudo-determinant of the covariance matrix.
-    pub fn pseudo_determinant(&self) -> f32 {
+    pub fn pseudo_determinant(&self) -> f64 {
         self.pseudo_determinant
     }
 }
 
-impl TryFrom<&[f32]> for CovMatrix {
+impl TryFrom<&[f64]> for CovMatrix {
     type Error = StatsError;
 
-    fn try_from(value: &[f32]) -> Result<Self, Self::Error> {
+    fn try_from(value: &[f64]) -> Result<Self, Self::Error> {
         Self::from_matrix(
             &DMatrix::from_diagonal(&DVector::from_iterator(value.len(), value.iter().copied()))
                 .as_view(),
@@ -201,18 +201,18 @@ impl TryFrom<&[f32]> for CovMatrix {
     }
 }
 
-impl<'a> TryFrom<&DMatrixView<'a, f32>> for CovMatrix {
+impl<'a> TryFrom<&DMatrixView<'a, f64>> for CovMatrix {
     type Error = StatsError;
 
-    fn try_from(value: &DMatrixView<'a, f32>) -> Result<Self, Self::Error> {
+    fn try_from(value: &DMatrixView<'a, f64>) -> Result<Self, Self::Error> {
         Self::from_matrix(value)
     }
 }
 
-impl Mul<f32> for CovMatrix {
+impl Mul<f64> for CovMatrix {
     type Output = CovMatrix;
 
-    fn mul(self, rhs: f32) -> Self::Output {
+    fn mul(self, rhs: f64) -> Self::Output {
         let dim = self.matrix.ncols() as i32;
 
         Self::Output {
@@ -224,7 +224,7 @@ impl Mul<f32> for CovMatrix {
     }
 }
 
-impl Mul<CovMatrix> for f32 {
+impl Mul<CovMatrix> for f64 {
     type Output = CovMatrix;
     fn mul(self, rhs: CovMatrix) -> Self::Output {
         let dim = rhs.matrix.ncols() as i32;
@@ -238,8 +238,8 @@ impl Mul<CovMatrix> for f32 {
     }
 }
 
-impl MulAssign<f32> for CovMatrix {
-    fn mul_assign(&mut self, rhs: f32) {
+impl MulAssign<f64> for CovMatrix {
+    fn mul_assign(&mut self, rhs: f64) {
         let dim = self.matrix.ncols() as i32;
 
         self.cholesky_ltm *= rhs.sqrt();
@@ -252,9 +252,9 @@ impl MulAssign<f32> for CovMatrix {
 /// Computes the unbiased covariance over two slices.
 ///
 /// The length of both iterators must be equal (panic).
-pub fn covariance<'a, I>(x: I, y: I) -> f32
+pub fn covariance<'a, I>(x: I, y: I) -> f64
 where
-    I: IntoIterator<Item = &'a f32>,
+    I: IntoIterator<Item = &'a f64>,
     <I as IntoIterator>::IntoIter: Clone,
 {
     let x_iter = x.into_iter();
@@ -262,22 +262,22 @@ where
 
     let length = x_iter.clone().fold(0, |acc, _| acc + 1);
 
-    let mu_x = x_iter.clone().sum::<f32>() / length as f32;
-    let mu_y = x_iter.clone().sum::<f32>() / length as f32;
+    let mu_x = x_iter.clone().sum::<f64>() / length as f64;
+    let mu_y = x_iter.clone().sum::<f64>() / length as f64;
 
     zip_eq(x_iter, y_iter)
         .map(|(val_x, val_y)| (mu_x - val_x) * (mu_y - val_y))
-        .sum::<f32>()
-        / (length - 1) as f32
+        .sum::<f64>()
+        / (length - 1) as f64
 }
 
 /// Computes the unbiased covariance over two slices with weights.
 ///
 /// The length of all three iterators must be equal (panic).
-pub fn covariance_with_weights<'a, IV, IW>(x: IV, y: IV, w: IW) -> f32
+pub fn covariance_with_weights<'a, IV, IW>(x: IV, y: IV, w: IW) -> f64
 where
-    IV: IntoIterator<Item = &'a f32>,
-    IW: IntoIterator<Item = &'a f32>,
+    IV: IntoIterator<Item = &'a f64>,
+    IW: IntoIterator<Item = &'a f64>,
     <IV as IntoIterator>::IntoIter: Clone,
     <IW as IntoIterator>::IntoIter: Clone,
 {
@@ -285,24 +285,24 @@ where
     let y_iter = y.into_iter();
     let w_iter = w.into_iter();
 
-    let wsum = w_iter.clone().sum::<f32>();
-    let wsumsq = w_iter.clone().map(|val_w| val_w.powi(2)).sum::<f32>();
+    let wsum = w_iter.clone().sum::<f64>();
+    let wsumsq = w_iter.clone().map(|val_w| val_w.powi(2)).sum::<f64>();
 
     let wfac = wsum - wsumsq / wsum;
 
     let mu_x = zip_eq(x_iter.clone(), w_iter.clone())
         .map(|(val_x, val_w)| val_x * val_w)
-        .sum::<f32>()
+        .sum::<f64>()
         / wsum;
 
     let mu_y = zip_eq(y_iter.clone(), w_iter.clone())
         .map(|(val_y, val_w)| val_y * val_w)
-        .sum::<f32>()
+        .sum::<f64>()
         / wsum;
 
     zip_eq(x_iter, zip_eq(y_iter, w_iter))
         .map(|(val_x, (val_y, val_w))| val_w * (mu_x - val_x) * (mu_y - val_y))
-        .sum::<f32>()
+        .sum::<f64>()
         / wfac
 }
 
@@ -330,7 +330,7 @@ mod tests {
                 &[1.0, 0.8, 1.1, 1.3, 0.9]
             ) - 19.523237)
                 .abs()
-                < f32::EPSILON
+                < f64::EPSILON
         );
     }
 
@@ -339,14 +339,14 @@ mod tests {
         let mut rng = Xoshiro256PlusPlus::seed_from_u64(1);
         let uniform = Uniform::new(-0.5, 0.5).unwrap();
 
-        let array: Matrix<f32, _, Dyn, _> =
-            Matrix::<f32, U3, Dyn, VecStorage<f32, U3, Dyn>>::from_iterator(
+        let array: Matrix<f64, _, Dyn, _> =
+            Matrix::<f64, U3, Dyn, VecStorage<f64, U3, Dyn>>::from_iterator(
                 10,
                 (0..30).map(|idx| {
                     if idx % 3 == 1 {
                         0.0
                     } else {
-                        (idx % 3) as f32 + rng.sample(uniform)
+                        (idx % 3) as f64 + rng.sample(uniform)
                     }
                 }),
             );
@@ -355,12 +355,12 @@ mod tests {
 
         let covmat = CovMatrix::from_vectors(array_view, None).unwrap();
 
-        assert!((covmat.cholesky_ltm[(0, 0)] - 0.40718567).abs() < f32::EPSILON);
-        assert!((covmat.cholesky_ltm[(2, 0)] - 0.07841061).abs() < f32::EPSILON);
+        assert!((covmat.cholesky_ltm[(0, 0)] - 0.40718567).abs() < f64::EPSILON);
+        assert!((covmat.cholesky_ltm[(2, 0)] - 0.07841061).abs() < f64::EPSILON);
 
-        assert!((covmat.inverse_matrix[(0, 0)] - 6.915894).abs() < f32::EPSILON);
-        assert!((covmat.inverse_matrix[(2, 0)] + 4.5933948).abs() < f32::EPSILON);
+        assert!((covmat.inverse_matrix[(0, 0)] - 6.915894).abs() < f64::EPSILON);
+        assert!((covmat.inverse_matrix[(2, 0)] + 4.5933948).abs() < f64::EPSILON);
 
-        assert!((covmat.pseudo_determinant() - 0.0069507807).abs() < f32::EPSILON);
+        assert!((covmat.pseudo_determinant() - 0.0069507807).abs() < f64::EPSILON);
     }
 }
